refactor(shared): rename foundation to shared
The `foundation/` directory held a single generic `Result<T,E>` helper.
Rename to `shared/` to match the Node/TS community convention (shared/
common/utils) and drop an unnecessary internal vocabulary.
