+++ src/shared/result.ts
